+++ src/main/java/com/solace/samples/features/others/RRGuaranteedReplier.java
@@ -190,7 +190,7 @@ public class RRGuaranteedReplier extends SampleApp {
                 case TIMES:
                     return (double)(leftOperand * rightOperand);
                 case DIVIDE:
-                    return (double)((double)leftOperand / (double)rightOperand);
+                    return (double)leftOperand / (double)rightOperand;
                 default:
                     throw new Exception("Unkown operation");
             }
